Rename algebraic-type-function-matching => algebraic-type-matching-void
